# res/swagger.json
{"schemes":["http","https"],"host":"localhost:8080","basePath":"/","paths":{"/student/register_student":{"post":{"parameters":[{"type":"object","x-go-name":"Student","description":"Student","name":"Student","in":"body","schema":{"type":"object","properties":{"first_name":{"type":"string"},"last_name":{"type":"string"}}}}],"operationId":"CreateStudentHandler","summary":"Create Student Handler","tags":["student"],"responses":{}}},"/student/update_student":{"put":{"parameters":[{"type":"object","x-go-name":"Student","description":"Student","name":"Student","in":"body","schema":{"type":"object","properties":{"first_name":{"type":"string"},"last_name":{"type":"string"}}}}],"operationId":"UpdateStudentHandler","summary":"Update Student Handler","tags":["student"],"responses":{}}},"/student/{student_id}":{"delete":{"parameters":[{"type":"string","x-go-name":"student_id","description":"student _ id","name":"student_id","in":"path","schema":{"type":""}}],"operationId":"DeleteStudentHandler","summary":"Delete Student Handler","tags":["student"],"responses":{}},"get":{"parameters":[{"type":"string","x-go-name":"student_id","description":"student _ id","name":"student_id","in":"path","schema":{"type":""}}],"operationId":"GetStudentHandler","summary":"Get Student Handler","tags":["student"],"responses":{}}},"/subject/{subject_id}":{"get":{"parameters":[{"type":"string","x-go-name":"subject_id","description":"subject _ id","name":"subject_id","in":"path","schema":{"type":""}}],"operationId":"GetStudentsInSubjectHandler","summary":"Get Students In Subject Handler","tags":["subject"],"responses":{}}},"/subject/{subject_id}/student/{student_id}":{"delete":{"parameters":[{"type":"string","x-go-name":"student_id","description":"student _ id","name":"student_id","in":"path","schema":{"type":""}},{"type":"string","x-go-name":"subject_id","description":"subject _ id","name":"subject_id","in":"path","schema":{"type":""}}],"operationId":"DeleteStudentSubjectHandler","summary":"Delete Student Subject Handler","tags":["subject"],"responses":{}},"post":{"parameters":[{"type":"string","x-go-name":"student_id","description":"student _ id","name":"student_id","in":"path","schema":{"type":""}},{"type":"string","x-go-name":"subject_id","description":"subject _ id","name":"subject_id","in":"path","schema":{"type":""}}],"operationId":"CreateStudentSubjectHandler","summary":"Create Student Subject Handler","tags":["subject"],"responses":{}}}},"swagger":"2.0"}